Major overhaul to control multiple bulbs async
The original code would send two requests to the tv for each Yeelight (every 0.05 seconds, so that would mean 60 requests / second for just 3 Yeelights).
This is now rewritten in such a way that the bulbs will be observers. Only two requests will be send to the TV (one only every 10 seconds, and the other every 0.05 seconds, So just 21 requests for any number of Yeelights). The response will be send to a infinite amount of bulbs that have subscribed themselves to be an observer.